split the CLI into modules and table-drive provider prompts
index.ts held the whole CLI, including nine provider setup flows that
differed only in their labels, env keys, and target lists. Those become a
descriptor table read by one collectProvider, and the commands, compose
rendering, docker invocation, and config handling move to their own
modules. A test pins the prompt sequence each provider kind produces.
