+++ apps/cli/src/commands/compose.ts
@@ -1,0 +1,8 @@
+import { type DirOption, resolveConfigDir } from "../config.js";
+import { assertDockerRunning, runDockerCompose } from "../docker.js";
+
+export async function runCompose(args: string[], options: DirOption): Promise<void> {
+	const configDir = await resolveConfigDir(options.dir);
+	assertDockerRunning();
+	await runDockerCompose(configDir, args);
+}
