+++ apps/cli/src/commands/init.ts
@@ -1,0 +1,451 @@
+import crypto from "node:crypto";
+import fs from "node:fs/promises";
+import path from "node:path";
+import * as p from "@clack/prompts";
+import { CLOUD_ENTRY_PRICE_USD } from "@workspace/config/plans";
+import { cloudPricingUrl } from "@workspace/config/referrals";
+import { formatScrapeTarget } from "@workspace/config/scrape-targets";
+import { parse as parseDotenv } from "dotenv";
+import pc from "picocolors";
+import { buildComposeYaml } from "../compose.js";
+import { CONFIG_HOME, type EnvMap, ensureDir, fileExists, type PostgresMode, writeConfigFiles } from "../config.js";
+import { assertDockerRunning, runDockerCompose, waitForHealthy } from "../docker.js";
+import {
+	collectCredentials,
+	collectProvider,
+	DEFAULT_SCRAPER_MODELS,
+	type DirectApiProvider,
+	finalizeScrapeTargets,
+	hasDirectApiConfigured,
+	PROVIDER_PROMPTS,
+	providerSpec,
+	type RecommendedScraper,
+} from "../providers.js";
+import { submitNewsletterSignup, trackCliEvent } from "../telemetry.js";
+import { assertNotCancelled, generateSecret, link, log, printBanner } from "../util.js";
+
+export type InitOptions = {
+	dev?: boolean;
+	dir?: string;
+	dockerDir?: string;
+};
+
+const DEFAULT_APP_NAME = "Elmo";
+const DEFAULT_APP_ICON = "/icons/elmo-icon.svg";
+const DEFAULT_APP_PORT = 1515;
+const LOCAL_DATABASE_URL = "postgres://postgres:postgres@postgres:5432/elmo";
+const TELEMETRY_DOC_URL = "https://elmohq.com/docs/developer-guide/telemetry";
+const CLOUD_PRICING_URL = cloudPricingUrl("cli");
+
+/**
+ * Guard an existing config from being replaced by accident. Returns the
+ * deployment id to carry over, so a re-run keeps reporting as the same install.
+ */
+async function confirmEnvOverwrite(configDir: string): Promise<string | undefined> {
+	const existingEnvPath = path.join(configDir, ".env");
+	if (!(await fileExists(existingEnvPath))) return undefined;
+
+	const contents = await fs.readFile(existingEnvPath, "utf8");
+	const isElmoEnv = contents.startsWith("# Rendered by elmo") || contents.startsWith("# Generated by elmo");
+
+	p.log.warn(
+		isElmoEnv
+			? `An existing Elmo config was found at ${configDir}.`
+			: `A .env file already exists in ${configDir} and was NOT created by Elmo.`,
+	);
+	const overwrite = await p.confirm({
+		message: isElmoEnv
+			? "Overwrite it with new values? Existing secrets (DATABASE_URL, API keys, etc.) will be replaced."
+			: "Overwrite the existing .env file? This cannot be undone.",
+		initialValue: false,
+	});
+	assertNotCancelled(overwrite);
+	if (!overwrite) {
+		p.cancel(
+			isElmoEnv
+				? "Setup cancelled. Use `elmo edit env` to change individual values."
+				: "Setup cancelled. Choose a different directory with --dir.",
+		);
+		process.exit(0);
+	}
+
+	return isElmoEnv ? parseDotenv(contents).DEPLOYMENT_ID : undefined;
+}
+
+async function resolveDockerDirInteractive(cwd: string): Promise<string> {
+	const inCwd = await fileExists(path.join(cwd, "Dockerfile"));
+	const inDockerDir = await fileExists(path.join(cwd, "docker", "Dockerfile"));
+	const defaultDir = inCwd ? "." : inDockerDir ? "docker" : ".";
+
+	const dir = await p.text({
+		message: "Path to docker directory (contains Dockerfile)",
+		defaultValue: defaultDir,
+	});
+	assertNotCancelled(dir);
+
+	const resolved = path.resolve(cwd, dir);
+	if (!(await fileExists(path.join(resolved, "Dockerfile")))) {
+		p.log.error(`Dockerfile not found in ${resolved}. Provide the directory that contains Dockerfile.`);
+		process.exit(1);
+	}
+
+	return resolved;
+}
+
+async function resolveDockerLocation(
+	options: InitOptions,
+	cwd: string,
+): Promise<{ dockerDir: string | undefined; repoRoot: string }> {
+	if (!options.dev) return { dockerDir: undefined, repoRoot: cwd };
+
+	if (!options.dockerDir) {
+		const dockerDir = await resolveDockerDirInteractive(cwd);
+		return { dockerDir, repoRoot: path.resolve(dockerDir, "..") };
+	}
+
+	const dockerDir = path.resolve(cwd, options.dockerDir);
+	if (!(await fileExists(path.join(dockerDir, "Dockerfile")))) {
+		p.log.error(`Dockerfile not found in ${dockerDir}`);
+		process.exit(1);
+	}
+	return { dockerDir, repoRoot: path.resolve(dockerDir, "..") };
+}
+
+function validatePort(value: string | undefined): string | undefined {
+	if (!value) return undefined;
+	const port = Number(value);
+	const valid = Number.isInteger(port) && port >= 1 && port <= 65535;
+	return valid ? undefined : "Must be an integer between 1 and 65535";
+}
+
+async function configureProvidersInteractive(env: EnvMap): Promise<"recommended" | "custom"> {
+	p.note(
+		[
+			"Elmo needs two kinds of providers:",
+			"",
+			pc.bold("1. A scraper") + " — to track ChatGPT and Google AI Mode (no public APIs):",
+			`     • ${pc.cyan("Cloro")}      — most reliable, every surface, ~$0.65/mo per prompt ($30/mo min)`,
+			`     • ${pc.cyan("BrightData")} — pay-as-you-go and cheaper, but slower, ~$0.45/mo per prompt`,
+			`     • ${pc.cyan("Oxylabs")}    — cheapest per run, no Gemini/Copilot, $49/mo min`,
+			`     • ${pc.cyan("Olostep")}    — premium, built for high volume, ~$2.25/mo per prompt`,
+			`     • ${pc.cyan("DataForSEO")} — pay-as-you-go, scrapers + direct APIs, ~$1.20/mo per prompt`,
+			"",
+			pc.bold("2. A direct LLM API") + " — for low-latency tasks (onboarding analysis, sentiment scoring,",
+			"   ad-hoc LLM calls). Required:",
+			`     • ${pc.cyan("OpenRouter")} — one key, all major models (recommended)`,
+			`     • ${pc.cyan("Anthropic / OpenAI / Mistral")} — direct provider keys`,
+			"",
+			"Pricing assumes Elmo's default cadence (5 runs/day × 2 surfaces).",
+		].join("\n"),
+		"AI providers",
+	);
+
+	const mode = await p.select({
+		message: "Setup mode",
+		options: [
+			{ value: "recommended" as const, label: "Recommended — one scraper + one direct API" },
+			{ value: "custom" as const, label: "Custom — pick each provider individually" },
+		],
+		initialValue: "recommended" as const,
+	});
+	assertNotCancelled(mode);
+
+	if (mode === "recommended") {
+		await configureProvidersRecommended(env);
+	} else {
+		await configureProvidersCustom(env);
+	}
+	return mode;
+}
+
+async function configureProvidersRecommended(env: EnvMap): Promise<void> {
+	const targets: string[] = [];
+
+	// ── Scraper ─────────────────────────────────────────────────────────────
+	const scraper = await p.select({
+		message: "Scraper (tracks ChatGPT + Google AI Mode)",
+		options: [
+			{ value: "cloro" as const, label: "Cloro — most reliable, every surface (~$0.65/mo per prompt, $30/mo min)" },
+			{ value: "brightdata" as const, label: "BrightData — pay-as-you-go, cheaper but slower (~$0.45/mo per prompt)" },
+			{ value: "oxylabs" as const, label: "Oxylabs — cheapest per run, no Gemini/Copilot ($49/mo min)" },
+			{ value: "olostep" as const, label: "Olostep — premium, built for high volume (~$2.25/mo per prompt)" },
+			{
+				value: "dataforseo" as const,
+				label: "DataForSEO — pay-as-you-go, scrapers + direct APIs (~$1.20/mo per prompt)",
+			},
+		],
+		initialValue: "cloro" as const,
+	});
+	assertNotCancelled(scraper);
+	await collectScraperKey(scraper, env);
+	for (const model of DEFAULT_SCRAPER_MODELS) {
+		targets.push(formatScrapeTarget({ model, provider: scraper, webSearch: true }));
+	}
+
+	// ── Direct API ──────────────────────────────────────────────────────────
+	const direct = await p.select({
+		message: "Direct LLM API (powers onboarding analysis + sentiment scoring)",
+		options: [
+			{ value: "openrouter" as const, label: "OpenRouter — one key, all major models (recommended)" },
+			{ value: "anthropic" as const, label: "Anthropic — direct Claude" },
+			{ value: "openai" as const, label: "OpenAI — direct GPT-* models" },
+			{ value: "mistral" as const, label: "Mistral — direct Mistral models" },
+		],
+		initialValue: "openrouter" as const,
+	});
+	assertNotCancelled(direct);
+	await collectDirectApiQuick(direct, env);
+
+	await finalizeScrapeTargets(env, targets, { skipEdit: true });
+}
+
+async function collectScraperKey(scraper: RecommendedScraper, env: EnvMap): Promise<void> {
+	const spec = providerSpec(scraper);
+	const url = spec.signup?.url;
+	if (url) p.log.info(`Sign up: ${link(pc.cyan(url), url)}`);
+	await collectCredentials(spec, env);
+}
+
+async function collectDirectApiQuick(kind: DirectApiProvider, env: EnvMap): Promise<void> {
+	await collectCredentials(providerSpec(kind), env);
+}
+
+async function configureProvidersCustom(env: EnvMap): Promise<void> {
+	const targets: string[] = [];
+
+	p.log.step(pc.bold("Step 1 of 2 — Direct LLM API (at least one is required)"));
+	while (!hasDirectApiConfigured(env)) {
+		for (const spec of PROVIDER_PROMPTS.filter((candidate) => candidate.kind === "direct")) {
+			await collectProvider(spec, env, targets);
+		}
+		if (!hasDirectApiConfigured(env)) {
+			p.log.warn(
+				"Onboarding analysis and other low-latency LLM tasks require a direct API. Configure at least one before continuing.",
+			);
+		}
+	}
+
+	p.log.step(pc.bold("Step 2 of 2 — Scrapers (optional, but needed to track ChatGPT / Google AI Mode)"));
+	for (const spec of PROVIDER_PROMPTS.filter((candidate) => candidate.kind === "scraper")) {
+		await collectProvider(spec, env, targets);
+	}
+
+	await finalizeScrapeTargets(env, targets);
+}
+
+async function doStart(configDir: string): Promise<void> {
+	assertDockerRunning();
+
+	log.step("Starting Docker Compose stack...");
+	await runDockerCompose(configDir, ["up", "-d"]);
+
+	const s = p.spinner();
+	s.start("Waiting for services to become healthy...");
+	const ok = await waitForHealthy(configDir, 180_000);
+	if (ok) {
+		s.stop("All services healthy!");
+	} else {
+		s.stop("Health check timed out.");
+		p.log.warn("Some services did not report healthy status.");
+	}
+
+	log.info("Examples:");
+	console.log(`  ${pc.bold("elmo compose logs -f")}`);
+	console.log(`  ${pc.bold("elmo compose logs -f web")}`);
+	console.log(`  ${pc.bold("elmo compose ps")}`);
+	console.log(`  ${pc.bold("elmo compose down")}`);
+}
+
+export async function runInit(options: InitOptions, version: string): Promise<void> {
+	printBanner();
+	p.intro(pc.bold("Setting up Elmo"));
+
+	// Said before the wizard asks for anything, not after: someone who would
+	// rather buy the managed version should find that out before they have gone
+	// and opened scraper accounts.
+	p.note(
+		[
+			"Self-hosting Elmo is free, and always will be. But it does mean",
+			"running Postgres and a worker, and opening your own scraper and",
+			"model provider accounts.",
+			"",
+			`Elmo Cloud is the same open-source product, hosted by us, from`,
+			`${pc.bold(`$${CLOUD_ENTRY_PRICE_USD}/month`)} — scraper and model access included, nothing to`,
+			"keep running.",
+			"",
+			`  ${link(pc.cyan(CLOUD_PRICING_URL), CLOUD_PRICING_URL)}`,
+		].join("\n"),
+		"Prefer not to self-host?",
+	);
+
+	const cwd = process.cwd();
+
+	// ── Resolve config directory ─────────────────────────────────────────
+	const configDir = options.dir ? path.resolve(cwd, options.dir) : CONFIG_HOME;
+
+	const preservedDeploymentId = await confirmEnvOverwrite(configDir);
+	const { dockerDir, repoRoot } = await resolveDockerLocation(options, cwd);
+
+	// ── Data stores ──────────────────────────────────────────────────────
+	const postgresMode = await p.select({
+		message: "PostgreSQL connection",
+		options: [
+			{
+				value: "docker" as const,
+				label: "Run Postgres in Docker",
+			},
+			{
+				value: "external" as const,
+				label: "Use existing Postgres (provide DATABASE_URL)",
+			},
+		],
+		initialValue: "docker" as PostgresMode,
+	});
+	assertNotCancelled(postgresMode);
+
+	const env: EnvMap = {};
+	env.DEPLOYMENT_MODE = "local";
+	env.VITE_DEPLOYMENT_MODE = "local";
+	env.DEPLOYMENT_ID = preservedDeploymentId ?? crypto.randomUUID();
+	env.BETTER_AUTH_SECRET = generateSecret();
+	// Standard base64 (not base64url): the app decodes this with Buffer.from(key,
+	// "base64") and requires exactly 32 bytes. Enables storing provider
+	// credentials encrypted in the database via the in-app Providers UI.
+	env.ELMO_ENCRYPTION_KEY = generateSecret(32, "base64");
+	env.APP_NAME = DEFAULT_APP_NAME;
+	env.APP_ICON = DEFAULT_APP_ICON;
+	env.VITE_APP_NAME = DEFAULT_APP_NAME;
+	env.VITE_APP_ICON = DEFAULT_APP_ICON;
+
+	if (postgresMode === "external") {
+		p.note("Must be an IPv4-compatible direct connection or database pooler.", "DATABASE_URL");
+		const url = await p.password({
+			message: "DATABASE_URL",
+			validate: (v) => (!v ? "Required" : undefined),
+		});
+		assertNotCancelled(url);
+		env.DATABASE_URL = url;
+	} else {
+		env.DATABASE_URL = LOCAL_DATABASE_URL;
+	}
+
+	// ── AI providers ─────────────────────────────────────────────────────
+	const setupMode = await configureProvidersInteractive(env);
+
+	// ── Telemetry ───────────────────────────────────────────────────────
+	p.note(
+		[
+			"Elmo is open source and maintained by a small team. Telemetry",
+			"from both the CLI and your local deployment (web + worker)",
+			"tells us things like which CLI versions are still in use, where",
+			"`elmo init` drops off, which providers people pick, and whether",
+			"new features actually get used. Without it we are flying blind",
+			"on what to fix or build next.",
+			"",
+			pc.bold("What we send:"),
+			"  • deployment ID (random UUID stored as DEPLOYMENT_ID in your .env)",
+			"  • CLI/app version, OS, arch, Node version, deployment mode",
+			"  • command/event names + non-secret options (e.g. postgres mode)",
+			"  • feature counts (prompts edited, brands created — never the names or text)",
+			"  • IP address (recorded on each event by PostHog, used for geolocation)",
+			"",
+			pc.bold("What we never send:"),
+			"  API keys, .env contents, brand names, prompt text, and scraped responses.",
+			"",
+			`Full breakdown: ${link(pc.cyan(TELEMETRY_DOC_URL), TELEMETRY_DOC_URL)}`,
+			"Toggle later by editing DISABLE_TELEMETRY in .env (`elmo edit env`).",
+		].join("\n"),
+		"Telemetry",
+	);
+
+	const telemetryEnabled = await p.confirm({
+		message: "Share telemetry?",
+		initialValue: true,
+	});
+	assertNotCancelled(telemetryEnabled);
+	if (!telemetryEnabled) {
+		env.DISABLE_TELEMETRY = "1";
+	}
+
+	// ── Product updates ─────────────────────────────────────────────────
+	const updatesEmail = await p.text({
+		message: "Enter your work email to receive product updates (optional)",
+		placeholder: "you@example.com",
+	});
+	const email = p.isCancel(updatesEmail) ? undefined : updatesEmail || undefined;
+
+	// ── Web app port ────────────────────────────────────────────────────
+	const portInput = await p.text({
+		message: "Web app port",
+		placeholder: String(DEFAULT_APP_PORT),
+		defaultValue: String(DEFAULT_APP_PORT),
+		validate: validatePort,
+	});
+	assertNotCancelled(portInput);
+	const port = Number(portInput);
+	env.APP_URL = `http://localhost:${port}`;
+	env.VITE_APP_URL = env.APP_URL;
+
+	// ── Write config ─────────────────────────────────────────────────────
+	const composeYaml = buildComposeYaml({
+		dev: Boolean(options.dev),
+		postgresMode,
+		repoRoot,
+		dockerDir,
+		port,
+		version,
+	});
+
+	await ensureDir(configDir);
+	await writeConfigFiles(configDir, {
+		env,
+		composeYaml,
+		postgresMode,
+		dev: Boolean(options.dev),
+		version,
+	});
+
+	p.log.success(`Config written to ${configDir}`);
+	p.log.warn("Your generated .env file contains secrets — do not commit it to version control.");
+
+	if (options.dev) {
+		p.log.info("Dev mode enabled. Run `elmo compose build` before starting.");
+	}
+
+	const shouldStart = await p.confirm({
+		message: "Start the stack now?",
+		initialValue: true,
+	});
+	assertNotCancelled(shouldStart);
+
+	if (shouldStart) {
+		await doStart(configDir);
+	} else {
+		p.log.info("You can start later with `elmo compose up -d`.");
+	}
+
+	// CLI telemetry — silently dropped if the user opted out above.
+	await trackCliEvent(configDir, "cli_init", {
+		version,
+		os: process.platform,
+		arch: process.arch,
+		node_version: process.version,
+		postgres_mode: postgresMode,
+		dev_mode: Boolean(options.dev),
+		setup_mode: setupMode,
+		has_scraper: Boolean(env.BRIGHTDATA_API_TOKEN || env.OLOSTEP_API_KEY || env.OXYLABS_USERNAME || env.CLORO_API_KEY),
+		has_direct_api: hasDirectApiConfigured(env),
+	});
+
+	// Newsletter signup is a separate, explicit opt-in and runs even when
+	// telemetry is disabled.
+	if (email) {
+		await submitNewsletterSignup(configDir, email);
+	}
+
+	p.log.message(
+		`If you find Elmo useful, star us on GitHub!\n  ${link(pc.cyan("https://github.com/elmohq/elmo"), "https://github.com/elmohq/elmo")}`,
+	);
+
+	p.outro(pc.green("Setup complete!"));
+}
